Ensure man page installation works with new huskymak.cfg file
In commit huskyproject/huskybse@ac231db, the huskymak.cfg file on which fidostat's makefile depends was rewritten. This removed the MANDIR variable that was used to specify where man pages are to be installed, and replaced it by separate MAN1DIR, MAN3DIR, and MAN5DIR variables. From now on, the man page of fidostat is installed into the directory specified by the MAN1DIR variable.